Fix compiler warning on Win64
BUG=None
LOG=N
TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/191153002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@19731 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-heap-profiler.cc
@@ -2524,19 +2524,19 @@ TEST(AddressToTraceMap) {
   CHECK_EQ(1, map.GetTraceNodeId(ToAddress(0x100)));
   CHECK_EQ(1, map.GetTraceNodeId(ToAddress(0x150)));
   CHECK_EQ(0, map.GetTraceNodeId(ToAddress(0x100 + 0x100)));
-  CHECK_EQ(1, map.size());
+  CHECK_EQ(1, static_cast<int>(map.size()));
 
   // [0x100, 0x200) -> 1, [0x200, 0x300) -> 2
   map.AddRange(ToAddress(0x200), 0x100, 2U);
   CHECK_EQ(2, map.GetTraceNodeId(ToAddress(0x2a0)));
-  CHECK_EQ(2, map.size());
+  CHECK_EQ(2, static_cast<int>(map.size()));
 
   // [0x100, 0x180) -> 1, [0x180, 0x280) -> 3, [0x280, 0x300) -> 2
   map.AddRange(ToAddress(0x180), 0x100, 3U);
   CHECK_EQ(1, map.GetTraceNodeId(ToAddress(0x17F)));
   CHECK_EQ(2, map.GetTraceNodeId(ToAddress(0x280)));
   CHECK_EQ(3, map.GetTraceNodeId(ToAddress(0x180)));
-  CHECK_EQ(3, map.size());
+  CHECK_EQ(3, static_cast<int>(map.size()));
 
   // [0x100, 0x180) -> 1, [0x180, 0x280) -> 3, [0x280, 0x300) -> 2,
   // [0x400, 0x500) -> 4
@@ -2547,22 +2547,22 @@ TEST(AddressToTraceMap) {
   CHECK_EQ(4, map.GetTraceNodeId(ToAddress(0x450)));
   CHECK_EQ(0, map.GetTraceNodeId(ToAddress(0x500)));
   CHECK_EQ(0, map.GetTraceNodeId(ToAddress(0x350)));
-  CHECK_EQ(4, map.size());
+  CHECK_EQ(4, static_cast<int>(map.size()));
 
   // [0x100, 0x180) -> 1, [0x180, 0x200) -> 3, [0x200, 0x600) -> 5
   map.AddRange(ToAddress(0x200), 0x400, 5U);
   CHECK_EQ(5, map.GetTraceNodeId(ToAddress(0x200)));
   CHECK_EQ(5, map.GetTraceNodeId(ToAddress(0x400)));
-  CHECK_EQ(3, map.size());
+  CHECK_EQ(3, static_cast<int>(map.size()));
 
   // [0x100, 0x180) -> 1, [0x180, 0x200) -> 7, [0x200, 0x600) ->5
   map.AddRange(ToAddress(0x180), 0x80, 6U);
   map.AddRange(ToAddress(0x180), 0x80, 7U);
   CHECK_EQ(7, map.GetTraceNodeId(ToAddress(0x180)));
   CHECK_EQ(5, map.GetTraceNodeId(ToAddress(0x200)));
-  CHECK_EQ(3, map.size());
+  CHECK_EQ(3, static_cast<int>(map.size()));
 
   map.Clear();
-  CHECK_EQ(0, map.size());
+  CHECK_EQ(0, static_cast<int>(map.size()));
   CHECK_EQ(0, map.GetTraceNodeId(ToAddress(0x400)));
 }
